move smart-on-fhir test; include qc in docker-compose-integration
